stock-agent/requirements.txt
# Stock-Agent Requirements

# 核心依赖
tushare>=1.2.89          # Tushare Pro 数据源（主）
akshare>=1.18.0          # AKShare 数据源（备用）
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# 技术指标
ta-lib>=0.4.0
# 或者纯 Python 实现
pandas-ta>=0.3.0

# 数据处理
requests>=2.31.0
aiohttp>=3.9.0

# 日志
loguru>=0.7.0

# 配置管理
python-dotenv>=1.0.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 可选：可视化
matplotlib>=3.8.0
plotly>=5.18.0

# 可选：机器学习
scikit-learn>=1.3.0
